[XrdSciTokens] Refactor scitoken helper classes into header
The various XrdSciTokens helper classes have been defined in-line
in the `XrdSciTokensAccess.cc` translation unit.  However, to do
standalone unit tests of these, refactor their definitions to be
in a standalone header.
