+++ src/XrdSciTokens/XrdSciTokensAccess.cc
@@ -24,6 +24,7 @@
 #include "picojson.h"
 
 #include "scitokens/scitokens.h"
+#include "XrdSciTokens/XrdSciTokensAccess.hh"
 #include "XrdSciTokens/XrdSciTokensHelper.hh"
 #include "XrdSciTokens/XrdSciTokensMon.hh"
 
@@ -31,6 +32,8 @@
 // linker definition and invoke directly.
 XrdVERSIONINFO(XrdAccAuthorizeObject, XrdAccSciTokens);
 XrdVERSIONINFO(XrdAccAuthorizeObjAdd, XrdAccSciTokens);
+
+XrdSciTokensHelper *SciTokensHelper = nullptr;
 
 namespace {
 
@@ -40,13 +43,6 @@ enum LogMask {
     Warning = 0x04,
     Error = 0x08,
     All = 0xff
-};
-
-enum IssuerAuthz {
-    Capability = 0x01,
-    Group = 0x02,
-    Mapping = 0x04,
-    Default = 0x07
 };
 
 std::string LogMaskToString(int mask) {
@@ -72,8 +68,6 @@ std::string LogMaskToString(int mask) {
     }
     return ss.str();
 }
-
-typedef std::vector<std::pair<Access_Operation, std::string>> AccessRulesRaw;
 
 inline uint64_t monotonic_time() {
   struct timespec tp;
@@ -224,62 +218,6 @@ void ParseCanonicalPaths(const std::string &path, std::vector<std::string> &resu
     } while (pos != std::string::npos);
 }
 
-struct MapRule
-{
-    MapRule(const std::string &sub,
-            const std::string &username,
-            const std::string &path_prefix,
-            const std::string &group,
-            const std::string &result)
-        : m_sub(sub),
-          m_username(username),
-          m_path_prefix(path_prefix),
-          m_group(group),
-          m_result(result)
-    {
-        //std::cerr << "Making a rule {sub=" << sub << ", username=" << username << ", path=" << path_prefix << ", group=" << group << ", result=" << name << "}" << std::endl;
-    }
-
-    const std::string match(const std::string &sub,
-                            const std::string &username,
-                            const std::string_view &req_path,
-                            const std::vector<std::string> &groups) const
-    {
-        if (!m_sub.empty() && sub != m_sub) {return "";}
-
-        if (!m_username.empty() && username != m_username) {return "";}
-
-        if (!m_path_prefix.empty() &&
-            strncmp(req_path.data(), m_path_prefix.c_str(), m_path_prefix.size()))
-        {
-            return "";
-        }
-
-        if (!m_group.empty()) {
-            for (const auto &group : groups) {
-                if (group == m_group)
-                    return m_result;
-            }
-            return "";
-        }
-        return m_result;
-    }
-
-    std::string m_sub;
-    std::string m_username;
-    std::string m_path_prefix;
-    std::string m_group;
-    std::string m_result;
-};
-
-// Control whether a given issuer is required for the paths it authorizes
-enum class AuthzSetting {
-    None, // Issuer's authorization is not necessary
-    Read, // Authorization from this issuer is necessary for reads.
-    Write, // Authorization from this issuer is necessary for writes.
-    All, // Authorization from this issuer is necessary for all operations.
-};
-
 struct IssuerConfig
 {
     IssuerConfig(const std::string &issuer_name,
@@ -358,149 +296,61 @@ protected:
 
 };
 
-// Given a list of access rules, this class determines whether a requested operation / path
-// is permitted by the access rules.
-class SubpathMatch final {
-public:
-    SubpathMatch() = default;
-    SubpathMatch(const AccessRulesRaw &rules)
-    : m_rules(rules)
-    {}
 
-    // Determine whether the known access rules permit the requested `oper` on `path`.
-    bool apply(Access_Operation oper, const std::string_view path) const {
-        auto is_subdirectory = [](const std::string_view& dir, const std::string_view& subdir) {
-            if (subdir.size() < dir.size())
-                return false;
-
-            if (subdir.compare(0, dir.size(), dir, 0, dir.size()) != 0)
-                return false;
-
-            return dir.size() == subdir.size() || subdir[dir.size()] == '/' || dir == "/";
-        };
-
-        for (const auto & rule : m_rules) {
-            // Skip rules that don't match the current operation
-            if (rule.first != oper)
-                continue;
-
-            // If the rule allows any path, allow the operation
-            if (rule.second == "/")
-                return true;
-
-            // Allow operation if path is a subdirectory of the rule's path
-            if (is_subdirectory(rule.second, path)) {
-                return true;
-            } else {
-                // Allow stat and mkdir of parent directories to comply with WLCG token specs
-                if (oper == AOP_Stat || oper == AOP_Mkdir)
-                if (is_subdirectory(path, rule.second))
-                    return true;
-            }
-        }
-        return false;
-    }
-
-    bool empty() const {return m_rules.empty();} // Returns true if there are no rules to match
-
-    std::string str() const {return AccessRuleStr(m_rules);} // Returns a human-friendly representation of the access rules
-
-    size_t size() const {return m_rules.size();} // Returns the count of rules
-private:
-
-    AccessRulesRaw m_rules;
-};
-
-class XrdAccRules
+void
+ParseTokenString(const std::string &param, XrdOucEnv *env, std::vector<std::string_view> &authz_list)
 {
-public:
-    XrdAccRules(uint64_t expiry_time, const std::string &username, const std::string &token_subject,
-        const std::string &issuer, const std::vector<MapRule> &rules, const std::vector<std::string> &groups,
-        uint32_t authz_strategy, AuthzSetting acceptable_authz) :
-        m_authz_strategy(authz_strategy),
-        m_acceptable_authz(acceptable_authz),
-        m_expiry_time(expiry_time),
-        m_username(username),
-        m_token_subject(token_subject),
-        m_issuer(issuer),
-        m_map_rules(rules),
-        m_groups(groups)
-    {}
-
-    ~XrdAccRules() {}
-
-    bool apply(Access_Operation oper, const std::string_view path) {
-        return m_matcher.apply(oper, path);
-    }
-
-    bool expired() const {return monotonic_time() > m_expiry_time;}
-
-    void parse(const AccessRulesRaw &rules) {
-        m_matcher = SubpathMatch(rules);
-    }
-
-    std::string get_username(const std::string_view &req_path) const
-    {
-        for (const auto &rule : m_map_rules) {
-            std::string name = rule.match(m_token_subject, m_username, req_path, m_groups);
-            if (!name.empty()) {
-                return name;
-            }
+    if (!env) {return;}
+    const char *authz = env->Get(param.c_str());
+    if (!authz) {return;}
+    std::string_view authz_view(authz);
+    size_t pos;
+    do {
+        // Note: this is more permissive than the plugin was previously.
+        // The prefix 'Bearer%20' used to be required as that's what HTTP
+        // required.  However, to make this more pleasant for XRootD protocol
+        // users, we now simply "handle" the prefix insterad of requiring it.
+        if (authz_view.substr(0, 9) == "Bearer%20") {
+            authz_view = authz_view.substr(9);
         }
-        return "";
-    }
+        pos = authz_view.find(",");
+        authz_list.push_back(authz_view.substr(0, pos));
+        authz_view = authz_view.substr(pos + 1);
+    } while (pos != std::string_view::npos);
+}
 
-    const std::string str() const
-    {
-        std::stringstream ss;
-        ss << "mapped_username=" << m_username << ", subject=" << m_token_subject
-           << ", issuer=" << m_issuer;
-        if (!m_groups.empty()) {
-            ss << ", groups=";
-            bool first=true;
-            for (const auto &group : m_groups) {
-                ss << (first ? "" : ",") << group;
-                first = false;
-            }
+} // namespace
+
+std::string
+SubpathMatch::str() const {
+    return AccessRuleStr(m_rules); // Returns a human-friendly representation of the access rules
+}
+
+// Convert a list of authorizations into a human-readable string.
+const std::string
+XrdAccRules::str() const
+{
+    std::stringstream ss;
+    ss << "mapped_username=" << m_username << ", subject=" << m_token_subject
+        << ", issuer=" << m_issuer;
+    if (!m_groups.empty()) {
+        ss << ", groups=";
+        bool first=true;
+        for (const auto &group : m_groups) {
+            ss << (first ? "" : ",") << group;
+            first = false;
         }
-        if (!m_matcher.empty()) {
-            ss << ", authorizations=" << m_matcher.str();
-        }
-        return ss.str();
     }
-
-
-        // Return the token's subject, an opaque unique string within the issuer's
-        // namespace.  It may or may not be related to the username one should
-        // use within the authorization framework.
-    const std::string & get_token_subject() const {return m_token_subject;}
-    const std::string & get_default_username() const {return m_username;}
-    const std::string & get_issuer() const {return m_issuer;}
-
-    uint32_t get_authz_strategy() const {return m_authz_strategy;}
-    bool acceptable_authz(Access_Operation oper) const {
-        if (m_acceptable_authz == AuthzSetting::All) return true;
-        if (m_acceptable_authz == AuthzSetting::None) return false;
-
-        bool is_read = oper == AOP_Read || oper == AOP_Readdir || oper == AOP_Stat;
-        if (is_read) return m_acceptable_authz == AuthzSetting::Read;
-        else return m_acceptable_authz == AuthzSetting::Write;
+    if (!m_matcher.empty()) {
+        ss << ", authorizations=" << m_matcher.str();
     }
+    return ss.str();
+}
 
-    size_t size() const {return m_matcher.size();}
-    const std::vector<std::string> &groups() const {return m_groups;}
-
-private:
-    const uint32_t m_authz_strategy;
-    const AuthzSetting m_acceptable_authz;
-    SubpathMatch m_matcher;
-    const uint64_t m_expiry_time{0};
-    const std::string m_username;
-    const std::string m_token_subject;
-    const std::string m_issuer;
-    const std::vector<MapRule> m_map_rules;
-    const std::vector<std::string> m_groups;
-};
+bool XrdAccRules::expired() const
+{
+    return monotonic_time() > m_expiry_time;
+}
 
 // Determine whether a list of authorizations contains at least one entry
 // from each of the applicable required issuers.
@@ -582,7 +432,6 @@ bool AuthorizesRequiredIssuers(Access_Operation client_oper, const std::string_v
 class XrdAccSciTokens;
 
 XrdAccSciTokens *accSciTokens = nullptr;
-XrdSciTokensHelper *SciTokensHelper = nullptr;
 
 class XrdAccSciTokens : public XrdAccAuthorize, public XrdSciTokensHelper,
                         public XrdSciTokensMon
@@ -1609,8 +1458,6 @@ void InitAccSciTokens(XrdSysLogger *lp, const char *cfn, const char *parm,
         SciTokensHelper = accSciTokens;
     } catch (std::exception &) {
     }
-}
-
 }
 
 extern "C" {
